delete duplicate files and reuse codes
