+++ dubbo-remoting/dubbo-remoting-api/src/main/java/org/apache/dubbo/remoting/Constants.java
@@ -85,14 +85,6 @@ public interface Constants {
 
     String CODEC_VERSION_KEY = "codec.version";
 
-    String PU_STATE = "PUState";
-
-    String PU_SERVER_KEY = "PUServer";
-
-    String PU_CLIENT_KEY = "PUClient";
-
-    String PU_TRANSPORTER_KEY = "PUTransporter";
-
     String SERVER_KEY = "server";
 
     String CLIENT_KEY = "client";
