--- dubbo-remoting/dubbo-remoting-netty4/src/main/java/org/apache/dubbo/remoting/transport/netty4/portunification/PuNettyChannel.java
@@ -1,214 +1,0 @@
-package org.apache.dubbo.remoting.transport.netty4.portunification;
-
-import org.apache.dubbo.common.URL;
-import org.apache.dubbo.common.logger.Logger;
-import org.apache.dubbo.common.logger.LoggerFactory;
-import org.apache.dubbo.remoting.ChannelHandler;
-import org.apache.dubbo.remoting.RemotingException;
-import org.apache.dubbo.remoting.transport.AbstractChannel;
-import org.apache.dubbo.remoting.utils.PayloadDropper;
-
-import io.netty.channel.Channel;
-import io.netty.channel.ChannelFuture;
-
-import java.net.InetSocketAddress;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.atomic.AtomicBoolean;
-
-import static org.apache.dubbo.common.constants.CommonConstants.DEFAULT_TIMEOUT;
-import static org.apache.dubbo.common.constants.CommonConstants.TIMEOUT_KEY;
-
-public class PuNettyChannel extends AbstractChannel {
-    private static final Logger logger = LoggerFactory.getLogger(PuNettyChannel.class);
-    /**
-     * the cache for netty channel and dubbo channel
-     */
-    private static final ConcurrentMap<Channel, PuNettyChannel> CHANNEL_MAP = new ConcurrentHashMap<>();
-    /**
-     * netty channel
-     */
-    private final Channel channel;
-
-    private final Map<String, Object> attributes = new ConcurrentHashMap<String, Object>();
-
-    private final AtomicBoolean active = new AtomicBoolean(false);
-
-    public PuNettyChannel(Channel channel, URL url, ChannelHandler handler) {
-        super(url, handler);
-        this.channel = channel;
-    }
-
-    static PuNettyChannel getOrAddChannel(Channel ch, URL url, ChannelHandler handler) {
-        if (ch == null) {
-            return null;
-        }
-        PuNettyChannel ret = CHANNEL_MAP.get(ch);
-        if (ret == null) {
-            PuNettyChannel nettyChannel = new PuNettyChannel(ch, url, handler);
-            if (ch.isActive()) {
-                nettyChannel.markActive(true);
-                ret = CHANNEL_MAP.putIfAbsent(ch, nettyChannel);
-            }
-            if (ret == null) {
-                ret = nettyChannel;
-            }
-        }
-        return ret;
-    }
-
-    static void removeChannelIfDisconnected(Channel ch) {
-        if (ch != null && !ch.isActive()) {
-            PuNettyChannel nettyChannel = CHANNEL_MAP.remove(ch);
-            if (nettyChannel != null) {
-                nettyChannel.markActive(false);
-            }
-        }
-    }
-
-    static void removeChannel(Channel ch) {
-        if (ch != null) {
-            PuNettyChannel nettyChannel = CHANNEL_MAP.remove(ch);
-            if (nettyChannel != null) {
-                nettyChannel.markActive(false);
-            }
-        }
-    }
-
-    public void markActive(boolean b) {
-        active.set(b);
-    }
-
-    @Override
-    public InetSocketAddress getLocalAddress() {
-        return (InetSocketAddress) channel.localAddress();
-    }
-
-    @Override
-    public InetSocketAddress getRemoteAddress() {
-        return (InetSocketAddress) channel.remoteAddress();
-    }
-
-    @Override
-    public boolean isConnected() {
-        return !isClosed() && active.get();
-    }
-
-    @Override
-    public boolean hasAttribute(String key) {
-        return attributes.containsKey(key);
-    }
-
-    @Override
-    public Object getAttribute(String key) {
-        return attributes.get(key);
-    }
-
-    @Override
-    public void setAttribute(String key, Object value) {
-        if (value == null) {
-            attributes.remove(key);
-        } else {
-            attributes.put(key, value);
-        }
-    }
-
-    @Override
-    public void removeAttribute(String key) {
-        attributes.remove(key);
-    }
-
-    @Override
-    public void send(Object message, boolean sent) throws RemotingException {
-        // whether the channel is closed
-        super.send(message, sent);
-
-        boolean success = true;
-        int timeout = 0;
-        try {
-            ChannelFuture future = channel.writeAndFlush(message);
-            if (sent) {
-                // wait timeout ms
-                timeout = getUrl().getPositiveParameter(TIMEOUT_KEY, DEFAULT_TIMEOUT);
-                success = future.await(timeout);
-            }
-            Throwable cause = future.cause();
-            if (cause != null) {
-                throw cause;
-            }
-        } catch (Throwable e) {
-            removeChannelIfDisconnected(channel);
-            throw new RemotingException(this, "Failed to send message " + PayloadDropper.getRequestWithoutData(message) + " to " + getRemoteAddress() + ", cause: " + e.getMessage(), e);
-        }
-        if (!success) {
-            throw new RemotingException(this, "Failed to send message " + PayloadDropper.getRequestWithoutData(message) + " to " + getRemoteAddress()
-                + "in timeout(" + timeout + "ms) limit");
-        }
-    }
-
-
-    @Override
-    public void close() {
-        try {
-            super.close();
-        } catch (Exception e) {
-            logger.warn(e.getMessage(), e);
-        }
-        try {
-            removeChannelIfDisconnected(channel);
-        } catch (Exception e) {
-            logger.warn(e.getMessage(), e);
-        }
-        try {
-            attributes.clear();
-        } catch (Exception e) {
-            logger.warn(e.getMessage(), e);
-        }
-        try {
-            if (logger.isInfoEnabled()) {
-                logger.info("Close netty channel " + channel);
-            }
-            channel.close();
-        } catch (Exception e) {
-            logger.warn(e.getMessage(), e);
-        }
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((channel == null) ? 0 : channel.hashCode());
-        return result;
-    }
-
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) {
-            return true;
-        }
-        if (obj == null) {
-            return false;
-        }
-
-        if (getClass() != obj.getClass()) {
-            return false;
-        }
-        PuNettyChannel other = (PuNettyChannel) obj;
-        if (channel == null) {
-            if (other.channel != null) {
-                return false;
-            }
-        } else if (!channel.equals(other.channel)) {
-            return false;
-        }
-        return true;
-    }
-
-    @Override
-    public String toString() {
-        return "PuNettyChannel [channel=" + channel + "]";
-    }
-}
